Typsecript done
